add MD5 integrity checking when Content-MD5 is present
1. checksum the incoming PUT data and raise BadRequest on mismatch
1. checksum the retrieved GET data and raise IOError on mismatch

Note, the 200 OK response is already in progress when the GET checksum
is performed, but we make an attempt to discover the error before the
EOF is reached so that the server might close the connection early
and let the client know something is wrong.  The exception should also
be logged.