add get_data toggle to methods to support HEAD request processing
allows resource to skip creation of data_generator when the caller
will not be iterating over the data results.

resources can still return a data_generator but should assume that the
caller may ignore it, so it will go straight to garbage collector
without necessarily forcing evaluation of the generator.