remove `Location` header from documented GET and HEAD responses [ci skip]
The previously documented behavior was not actually implemented. A
closer reading of the HTTP specs only defines non-redirect behavior
for `Location` in `201 Created` PUT and POST responses and it seems
tooling may do unintended things if we try to use it as extra
information in a `200 OK` response.

If we want to support discovery of the versioned object currently
served from an unversioned object name resource, we need to consider
one of these alternatives instead:

  - Use a `302 Found` response to redirect to the current versioned URL, adding another round trip HTTP request.
  - Add a custom `X-Hatrac-Location` header or similar extension without core spec implications.