fix internal error in repeat PUT to namespace, simplify API and docs
There was some broken attempt at allowing idempotent namespace
creation by repeating a PUT action.  It seems simpler to just
return 409 Conflict on the second attempt and let the client
figure out that it exists either by interpreting the error text
or by doing a HEAD/GET and seeing that it exists.