also add `pid` column tracking parent ID on namespaces and objects
The new `pid` combined with `ancestors` can replace all the expensive
regular-expression based tests previously used to enumerate children
and descendents, an approach that was scaling poorly with table size.

This patch, like the `ancestors` patch, provides an auto-upgrade
mechanism to alter an existing Hatrac deployment in-place at service
startup.