fix regressions in non-object-content JSON responses; expand tests to cover
A previous refactoring was incomplete. Code that used to always
accumulate data generator output into a single response string was
previously removed, but resources that used to return a list of
strings were not adjusted to return a string or string sequence
generator. The effect was that a Python list of JSON was being
implicitly serialized as the response rather than the JSON content
itself.

Added JSON deserialization to rest-smoketest to check response content
on every response with Content-Type application/json. This would have
caught these errors...