make deploy_db and schema_upgrade idempotent and run both in re-deploy
The hatrac-deploy will stop spewing errors when run on an existing
deploy, and can incrementally upgrade the existing deploy.

Also had to fix some lingering assumptions about old web.db backend to
make schema_upgrade work with the current psycopg2 backend.

This is the first step in preparing a new metadata feature. Some of the
commits in this series may create errors. The final PR will be for the
complete set of changes...