add optional parents=true parameter to create namespaces, objects, or uploads
- `PUT  /path/child?parents=true`
- `POST /path/child;upload?parents=true`

This option requests the server to recursively try to fill in missing
ancestors in the path, as if the client had individually requested
each one to be created as separate requests.  I.e. the client must be
authorized to create those namespaces and the new namespaces (if any)
will be owned by that client.

The default remains to raise a `404 Not Found` error if the parent
path does not already exist.