fix potential bug for HEAD requests
Data generator should be empty byte sequence not `None` value. This
passed tests but may be the cause of some non-deterministic errors
seen on older systems, if web.py serializes the `None` and rewrites
the Content-Length response header with the value 4, replacing
the correct value we retrieve from the object registry.