add optional query parameter syntax to all URL dispatch rules
Even when we don't expect query paramters, change the URL patterns to
accept and supply the `?...` suffix and pass this as an extra
`querystr` parameter to each request handler method.  The method can
ignore or interpret this query string as it sees fit.

This acts more like other web servers and allows clients to play some
games with appending query parameters for application state tracking
and debugging w/o worrying about Hatrac-specific restrictions.