adjust documentation for namespace subtree-owner ACL to match implementation
This is arguably the lazy way out, but it seems less risky to adjust
the documentation than to change the ACL enforcement behavior.

The existing enforcement is less permissive than implied by the
previous documentation. If we adjust the implementation, existing
deployments might start to permit writes that are currently forbidden,
and this might be an unwelcome surprise to a service operator who
expects the existing behavior.

On the other hand, a site can extend the namespace owner ACL with
the same elements in the subtree-owner ACL in order to get that
behavior, if actually desired.