wip snapshot: add version.aux column for use by backends
This is an untested, minimal prototype to add JSON aux storage for
object versions.

1. Deploy version.aux column (new deploys and schema upgrades)
2. Load verison.aux column -> HatracObjectVersion.aux field
3. Thread aux field as keyword arg to backend storage APIs
  - storage.get_version_content_range(..., aux={})
  - storage.get_version_content(..., aux={})
  - storage.delete(..., aux={})

This is for a minimal prototype where a storage backend MAY use aux
content provisioned by an out of band means. The first use case is for
a migration tool to prepare a hatrac DB where each version record
retains an existing version.version text key to use in URLs and adds a
separate version.aux['version'] or similar text key to use when
addressing a backend storage object.

If we wish to support distinct storage backend version IDs for new
objects, we may need further internal API refactoring to thread the
aux content between the storage backend and the directory backend.