refactor RestException (wip)
- rebase on werkzeug.exceptions.HTTPException
- refactor text vs HTML content negotiation into base class
- add easier configuration for customizing error templates

The existing code had hacked in support for custom templates
in the hatrac_config.json like:

  "403_html": "...%(message)s...",

While keeping backwards-compatible support for those, add a
more flexible config option:

  "error_templates": {
    code: {
      content_type: template
    }
  }

where code can be a status code like "401" or the string "default" to
set a generic template for all error codes not explicitly configured.
To allow this, a template key "title" is also supported, and the new
default behavior is as if this configuration exists:

  "error_templates": {
    "default": {
      "text/plain": "%(description)s",
      "text/html": "<html><body><h1>%(title)s</h1><p>%(description)s</p></body></html>"
    }
  }

the template key "message" is retained as a backwards-compatible alias
to "description".

This is an initial sketch and needs more testing and probably cleanup.