add "overlay" backend storage
This is configured as:

  "storage_backend": "overlay",
  "overlay_backends": [
    {
      "storage_backend": <primary backend type>,
      <primary config options>...
    },
    {
      "storage_backend": <secondary backend type>,
      <secondary config options>...
    }
  ]

All writes go to the primary (first) storage backend, i.e.
new object-version creation/deletion or upload job actions.

Reads of existing object-version search down the list in
order and may be served from secondary backend(s).

This is useful for TEST servers, e.g. where you might clone
a production hatrac DB and provision a primary filesystem
backend on top of a read-only view of the production s3
or networked filesystem backend.

For consistency, it is recommended that you prune upload
job content from the DB when restoring a clone into the
test server, i.e.

  DELETE FROM chunk;
  DELETE FROM upload;

Such existing jobs would be unusable in the test server
since all stages of stateful upload jobs must be completed
in the same primary backend.