separate final overlay backend from search loop
Run the search loop on backends[0:-1] (all but last) to search
for the object version and eat an ObjectVersionMissing exception.
We force a search_slice to make sure the s3 backend will make
a synchronous object_get and detect missing objects.

If we finish the loop, run the final backends[-1] call
separately with the actual client-specified get_slice, which
may be None. This backend, if using s3, may use signed urls.

This optimizes an expected common test scenario with a local
primary filesystem overlay on a shared s3 bucket. The bucket
access can still use signed urls as desired, with only a slight
change in how some inconsistency error scenarios might be
surfaced.