add optional boolean parameter "unquote_object_keys" for s3 bucket configs
The optional parameter "unquote_object_keys" defaults to false
and uses the same quoted object names in the underlying S3 object
key as in the DB and the raw hatrac rest URLs.

For backwards compatibility to older deployments, setting this
true will unquote the object names embedded in the underlying
S3 object key, to follow the scheme from the legacy hatrac
codebase prior to the web.py -> flask refactoring.

{
    "storage_backend": "amazons3",
    ...
    "s3_config": {
        ...
        "buckets": {
            "/": {
                "bucket_name": "<my bucket name>",
                ...
                "unquote_object_keys": true,
                ...
            }
        }
    }
}