fix off-by-one buglet in amazons3 backend module
The internal logic uses python slices with half-open [start,end)
semantics but needs to translate to HTTP Range [start,end] closed
intervals.

This bug did not seem to impact actual hatrac clients, due to
other layers of the web framework trimming the extra byte we received
from S3. But, it can interfere with other direct use of the backend
API, such as in migration tools.